pyproject.toml
[project]
name = "agentic-rag-pro"
version = "0.1.0"
description = "Agentic RAG app with Streamlit, LangChain, ChromaDB and multi-LLM routing"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "chromadb>=0.5.0",
    "langchain>=0.3.0",
    "langchain-chroma>=0.2.0",
    "langchain-community>=0.3.0",
    "langchain-core>=0.3.0",
    "langchain-google-genai>=4.2.1",
    "langchain-huggingface>=0.1.2",
    "langchain-openai>=0.3.0",
    "langchain-text-splitters>=1.1.1",
    "pypdf>=4.2.0",
    "pypdf2>=3.0.1",
    "pytest>=8.0.0",
    "python-dotenv>=1.0.1",
    "sentence-transformers>=3.0.0",
    "streamlit>=1.44.0",
]
